test: split registry context tests
