+++ crates/runner/mitm-addon/tests/test_registry_context.py
@@ -1,97 +1,11 @@
-"""Tests for registry VM lookup and compiled context behavior."""
+"""Tests for registry VM lookup and public compiled context behavior."""
 
 import json
-import os
 from unittest.mock import MagicMock, patch
-
-import pytest
 
 import matching
 import registry
-from tests.registry_helpers import pin_mtime, write_firewall_registry
-
-
-def _write_builtin_firewall_registry(
-    path,
-    *,
-    run_id: str,
-    name: str,
-    base_url_vars: dict[str, str],
-) -> None:
-    path.write_text(
-        json.dumps(
-            {
-                "vms": {
-                    "10.200.0.1": {
-                        "runId": run_id,
-                        "firewalls": [
-                            {
-                                "kind": "builtin",
-                                "name": name,
-                                "baseUrlVars": base_url_vars,
-                            }
-                        ],
-                    }
-                },
-                "updatedAt": 0,
-            }
-        )
-    )
-
-
-def _install_test_builtin_firewall(monkeypatch, *, name: str, base: str) -> None:
-    monkeypatch.setattr(
-        registry,
-        "BUILTIN_FIREWALLS",
-        {
-            name: {
-                "name": name,
-                "apis": [
-                    {
-                        "base": base,
-                        "auth": {"headers": {"Authorization": "Bearer ${{ secrets.API_TOKEN }}"}},
-                        "permissions": [],
-                    }
-                ],
-            }
-        },
-    )
-
-
-def _write_multi_vm_registry(path, vms: dict) -> None:
-    path.write_text(json.dumps({"vms": vms, "updatedAt": 0}))
-
-
-def _builtin_vm(run_id: str, name: str, base_url_vars: dict[str, str] | None = None) -> dict:
-    entry: dict[str, object] = {"kind": "builtin", "name": name}
-    if base_url_vars is not None:
-        entry["baseUrlVars"] = base_url_vars
-    return {"runId": run_id, "firewalls": [entry]}
-
-
-def _inline_vm(run_id: str) -> dict:
-    return {
-        "runId": run_id,
-        "firewalls": [
-            {
-                "kind": "inline",
-                "firewall": {
-                    "name": "example",
-                    "apis": [
-                        {
-                            "base": "https://api.example.com",
-                            "auth": {"headers": {"Authorization": "Bearer token"}},
-                            "permissions": [{"name": "read", "rules": ["GET /items"]}],
-                        }
-                    ],
-                },
-            }
-        ],
-    }
-
-
-def _first_firewall_core(compiled_firewalls: matching.CompiledFirewallSet):
-    return compiled_firewalls.firewalls[0].core
+from tests.registry_helpers import write_firewall_registry
 
 
 class TestGetVmInfo:
@@ -183,937 +97,6 @@ class TestGetVmContext:
         assert isinstance(result, matching.FirewallAllow)
         assert result.api_entry is vm_info["firewalls"][0]["apis"][0]
 
-    def test_builtin_firewall_entry_resolves_from_catalog(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-github",
-                            "firewalls": [{"kind": "builtin", "name": "github"}],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["name"] == "github"
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == "https://api.github.com"
-        assert vm_info["firewalls"][0]["apis"][0]["id"] == "run-github:0"
-
-    def test_repeated_builtin_firewall_refs_share_core_but_keep_vm_api_ids(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_multi_vm_registry(
-            path,
-            {
-                "10.200.0.1": _builtin_vm("run-github-a", "github"),
-                "10.200.0.2": _builtin_vm("run-github-b", "github"),
-            },
-        )
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is _first_firewall_core(second_compiled)
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://api.github.com/repos/vm0-ai/vm0",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://api.github.com/repos/vm0-ai/vm0",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_vm_info["firewalls"][0]["apis"][0]
-        assert second_result.api_entry is second_vm_info["firewalls"][0]["apis"][0]
-        assert first_result.api_entry["id"] == "run-github-a:0"
-        assert second_result.api_entry["id"] == "run-github-b:0"
-
-    def test_builtin_firewall_refs_share_static_payload_but_keep_vm_api_shells(
-        self, tmp_path, monkeypatch
-    ):
-        auth = {"headers": {"Authorization": "Bearer ${{ secrets.API_TOKEN }}"}}
-        permissions = [{"name": "read-items", "rules": ["GET /items"]}]
-        monkeypatch.setattr(
-            registry,
-            "BUILTIN_FIREWALLS",
-            {
-                "large": {
-                    "name": "large",
-                    "apis": [
-                        {
-                            "base": "https://api.example.com",
-                            "auth": auth,
-                            "permissions": permissions,
-                        },
-                        {
-                            "base": "https://upload.example.com",
-                            "auth": auth,
-                            "permissions": permissions,
-                        },
-                    ],
-                }
-            },
-        )
-        path = tmp_path / "registry.json"
-        _write_multi_vm_registry(
-            path,
-            {
-                "10.200.0.1": _builtin_vm("run-large-a", "large"),
-                "10.200.0.2": _builtin_vm("run-large-b", "large"),
-            },
-        )
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is _first_firewall_core(second_compiled)
-
-        first_firewall = first_vm_info["firewalls"][0]
-        second_firewall = second_vm_info["firewalls"][0]
-        assert first_firewall is not second_firewall
-        assert first_firewall["apis"] is not second_firewall["apis"]
-
-        first_api = first_firewall["apis"][0]
-        second_api = second_firewall["apis"][0]
-        assert first_api is not second_api
-        assert first_api["id"] == "run-large-a:0"
-        assert second_api["id"] == "run-large-b:0"
-        assert first_api["permissions"] is permissions
-        assert second_api["permissions"] is permissions
-        assert first_api["auth"] is auth
-        assert second_api["auth"] is auth
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_api
-        assert second_result.api_entry is second_api
-
-    def test_builtin_firewall_entry_resolves_dynamic_base_url_vars(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-zendesk",
-                            "firewalls": [
-                                {
-                                    "kind": "builtin",
-                                    "name": "zendesk",
-                                    "baseUrlVars": {"ZENDESK_SUBDOMAIN": "acme"},
-                                }
-                            ],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == "https://acme.zendesk.com"
-
-    def test_builtin_refs_with_different_base_url_vars_do_not_share_core(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_multi_vm_registry(
-            path,
-            {
-                "10.200.0.1": _builtin_vm(
-                    "run-zendesk-a",
-                    "zendesk",
-                    {"ZENDESK_SUBDOMAIN": "acme"},
-                ),
-                "10.200.0.2": _builtin_vm(
-                    "run-zendesk-b",
-                    "zendesk",
-                    {"ZENDESK_SUBDOMAIN": "beta"},
-                ),
-            },
-        )
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-        assert first_vm_info["firewalls"][0]["apis"][0]["base"] == "https://acme.zendesk.com"
-        assert second_vm_info["firewalls"][0]["apis"][0]["base"] == "https://beta.zendesk.com"
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://acme.zendesk.com/api/v2/tickets",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://beta.zendesk.com/api/v2/tickets",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_vm_info["firewalls"][0]["apis"][0]
-        assert second_result.api_entry is second_vm_info["firewalls"][0]["apis"][0]
-
-    def test_builtin_compile_cache_is_scoped_to_registry_path(self, tmp_path):
-        path_a = tmp_path / "registry-a.json"
-        path_b = tmp_path / "registry-b.json"
-        data = {"10.200.0.1": _builtin_vm("run-github", "github")}
-        _write_multi_vm_registry(path_a, data)
-        _write_multi_vm_registry(path_b, data)
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path_a))
-        second_context = registry.get_vm_context("10.200.0.1", str(path_b))
-
-        assert first_context is not None
-        assert second_context is not None
-        _, first_compiled, _ = first_context
-        _, second_compiled, _ = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-
-    def test_invalid_builtin_entry_does_not_poison_valid_vm_core_cache(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_multi_vm_registry(
-            path,
-            {
-                "10.200.0.1": _builtin_vm("run-github", "github"),
-                "10.200.0.2": _builtin_vm("run-missing", "missing-firewall"),
-            },
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            valid_context = registry.get_vm_context("10.200.0.1", str(path))
-            invalid_context = registry.get_vm_context("10.200.0.2", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert valid_context is not None
-        assert invalid_context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert state.invalid_vms["10.200.0.2"].reason == "invalid_firewalls"
-        _, compiled_firewalls, compiled_policies = valid_context
-        assert compiled_firewalls is not None
-        result = matching.match_compiled_firewall_request(
-            "https://api.github.com/repos/vm0-ai/vm0",
-            "GET",
-            compiled_firewalls,
-            compiled_policies,
-        )
-        assert isinstance(result, matching.FirewallAllow)
-
-    def test_builtin_core_cache_prunes_inactive_keys_on_registry_reload(
-        self, tmp_path, monkeypatch
-    ):
-        monkeypatch.setattr(
-            registry,
-            "BUILTIN_FIREWALLS",
-            {
-                "cache-a": {
-                    "name": "cache-a",
-                    "apis": [
-                        {
-                            "base": "https://api-a.example.com",
-                            "auth": {"headers": {"Authorization": "Bearer token"}},
-                            "permissions": [],
-                        }
-                    ],
-                },
-                "cache-b": {
-                    "name": "cache-b",
-                    "apis": [
-                        {
-                            "base": "https://api-b.example.com",
-                            "auth": {"headers": {"Authorization": "Bearer token"}},
-                            "permissions": [],
-                        }
-                    ],
-                },
-            },
-        )
-        path = tmp_path / "registry.json"
-        _write_multi_vm_registry(path, {"10.200.0.1": _builtin_vm("run-cache-a", "cache-a")})
-        os.utime(path, ns=(1_700_000_000_000_000_000, 1_700_000_000_000_000_000))
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert first_context is not None
-        assert len(registry._registry_state.builtin_firewall_core_cache) == 1
-        first_cache_key = next(iter(registry._registry_state.builtin_firewall_core_cache))
-        assert first_cache_key[0] == "cache-a"
-
-        _write_multi_vm_registry(path, {"10.200.0.1": _builtin_vm("run-cache-b", "cache-b")})
-        os.utime(path, ns=(1_700_000_000_000_000_001, 1_700_000_000_000_000_001))
-        second_context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert second_context is not None
-        assert len(registry._registry_state.builtin_firewall_core_cache) == 1
-        second_cache_key = next(iter(registry._registry_state.builtin_firewall_core_cache))
-        assert second_cache_key[0] == "cache-b"
-
-    def test_inline_firewalls_do_not_share_compiled_core(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_multi_vm_registry(
-            path,
-            {
-                "10.200.0.1": _inline_vm("run-inline-a"),
-                "10.200.0.2": _inline_vm("run-inline-b"),
-            },
-        )
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        second_context = registry.get_vm_context("10.200.0.2", str(path))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, first_policies = first_context
-        second_vm_info, second_compiled, second_policies = second_context
-        assert first_compiled is not None
-        assert second_compiled is not None
-        assert _first_firewall_core(first_compiled) is not _first_firewall_core(second_compiled)
-
-        first_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            first_compiled,
-            first_policies,
-        )
-        second_result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            second_compiled,
-            second_policies,
-        )
-
-        assert isinstance(first_result, matching.FirewallAllow)
-        assert isinstance(second_result, matching.FirewallAllow)
-        assert first_result.api_entry is first_vm_info["firewalls"][0]["apis"][0]
-        assert second_result.api_entry is second_vm_info["firewalls"][0]["apis"][0]
-
-    def test_builtin_fixed_provider_suffix_rejects_authority_escape(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-shopify",
-            name="shopify",
-            base_url_vars={"SHOPIFY_SHOP": "attacker.example:443/capture"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "SHOPIFY_SHOP"' in invalid_vm.message
-
-    def test_builtin_fixed_provider_suffix_rejects_encoded_structure(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-shopify",
-            name="shopify",
-            base_url_vars={"SHOPIFY_SHOP": "attacker.example%3A443%2Fcapture"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "SHOPIFY_SHOP"' in invalid_vm.message
-
-    def test_builtin_fixed_provider_suffix_accepts_multi_label_fragment(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-snowflake",
-            name="snowflake",
-            base_url_vars={"SNOWFLAKE_ACCOUNT": "xy12345.us-east-1.aws"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == (
-            "https://xy12345.us-east-1.aws.snowflakecomputing.com/api"
-        )
-
-    def test_builtin_fixed_provider_suffix_rejects_path_injection(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-snowflake",
-            name="snowflake",
-            base_url_vars={"SNOWFLAKE_ACCOUNT": "xy12345.us-east-1.aws/capture"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "SNOWFLAKE_ACCOUNT"' in invalid_vm.message
-
-    def test_builtin_whole_authority_accepts_host_value(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-jira",
-            name="jira",
-            base_url_vars={"JIRA_DOMAIN": "acme.atlassian.net"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == "https://acme.atlassian.net"
-
-    def test_builtin_whole_authority_rejects_path_injection(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-jira",
-            name="jira",
-            base_url_vars={"JIRA_DOMAIN": "attacker.example/capture"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "JIRA_DOMAIN"' in invalid_vm.message
-
-    def test_builtin_base_url_var_rejects_firewall_parameter_syntax(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-strapi",
-            name="strapi",
-            base_url_vars={"STRAPI_BASE_URL": "https://{host}.example.test"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "STRAPI_BASE_URL"' in invalid_vm.message
-
-    def test_builtin_base_url_var_rejects_unicode_whitespace(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-strapi",
-            name="strapi",
-            base_url_vars={"STRAPI_BASE_URL": "https://strapi.example.test/work\u00a0flows"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "STRAPI_BASE_URL"' in invalid_vm.message
-
-    def test_builtin_base_url_prefix_preserves_fixed_path_suffix(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-n8n",
-            name="n8n",
-            base_url_vars={"N8N_BASE_URL": "https://n8n.example.test/workflows"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == (
-            "https://n8n.example.test/workflows/api/v1"
-        )
-
-    def test_builtin_base_url_prefix_accepts_encoded_path_separator(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-n8n",
-            name="n8n",
-            base_url_vars={"N8N_BASE_URL": "https://n8n.example.test/work%2fflows"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == (
-            "https://n8n.example.test/work%2fflows/api/v1"
-        )
-
-    def test_builtin_base_url_prefix_rejects_path_dot_segments(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-n8n",
-            name="n8n",
-            base_url_vars={"N8N_BASE_URL": "https://n8n.example.test/workflows/.."},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "N8N_BASE_URL"' in invalid_vm.message
-
-    def test_builtin_base_url_prefix_rejects_encoded_path_dot_segments(self, tmp_path):
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-n8n",
-            name="n8n",
-            base_url_vars={"N8N_BASE_URL": "https://n8n.example.test/workflows/%2e%2e"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "N8N_BASE_URL"' in invalid_vm.message
-
-    def test_builtin_path_segment_var_accepts_fixed_suffix(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "acme:prod"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == (
-            "https://api.example.test/accounts/acme:prod/v1"
-        )
-
-    def test_builtin_path_segment_var_rejects_path_injection(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "acme/../admin"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "TENANT"' in invalid_vm.message
-
-    def test_builtin_path_segment_var_rejects_encoded_path_separator(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "acme%252fprod"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "TENANT"' in invalid_vm.message
-
-    def test_builtin_path_segment_var_rejects_dot_segment(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "%2e%2e"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "TENANT"' in invalid_vm.message
-
-    def test_builtin_path_segment_var_rejects_path_parameter_dot_segment(
-        self, tmp_path, monkeypatch
-    ):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "..;type=folder"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "TENANT"' in invalid_vm.message
-
-    def test_builtin_path_segment_var_rejects_nested_encoded_dot_segment(
-        self, tmp_path, monkeypatch
-    ):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "%252e%252e"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "TENANT"' in invalid_vm.message
-
-    def test_builtin_path_segment_var_rejects_compatibility_dot_segment(
-        self, tmp_path, monkeypatch
-    ):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="tenant-path",
-            base="https://api.example.test/accounts/${{ vars.TENANT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-tenant-path",
-            name="tenant-path",
-            base_url_vars={"TENANT": "\uff0e\uff0e"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "TENANT"' in invalid_vm.message
-
-    def test_builtin_path_var_allows_dot_outside_dot_segment(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="path-fragment",
-            base="https://api.example.test/v${{ vars.VERSION }}",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-path-fragment",
-            name="path-fragment",
-            base_url_vars={"VERSION": "%2e1"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == ("https://api.example.test/v%2e1")
-
-    def test_builtin_path_vars_reject_combined_dot_segment(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="multi-path-segment",
-            base="https://api.example.test/accounts/${{ vars.A }}${{ vars.B }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-multi-path-segment",
-            name="multi-path-segment",
-            base_url_vars={"A": ".", "B": "."},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert "resolved base URL has unsafe path segments" in invalid_vm.message
-
-    def test_builtin_path_vars_accept_combined_safe_segment(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="multi-path-segment",
-            base="https://api.example.test/accounts/${{ vars.A }}${{ vars.B }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-multi-path-segment",
-            name="multi-path-segment",
-            base_url_vars={"A": "v", "B": "1"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == (
-            "https://api.example.test/accounts/v1/v1"
-        )
-
-    def test_builtin_port_var_accepts_numeric_port(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="port-api",
-            base="https://api.example.test:${{ vars.API_PORT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-port-api",
-            name="port-api",
-            base_url_vars={"API_PORT": "8443"},
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == ("https://api.example.test:8443/v1")
-
-    def test_builtin_port_var_rejects_non_numeric_port(self, tmp_path, monkeypatch):
-        _install_test_builtin_firewall(
-            monkeypatch,
-            name="port-api",
-            base="https://api.example.test:${{ vars.API_PORT }}/v1",
-        )
-        path = tmp_path / "registry.json"
-        _write_builtin_firewall_registry(
-            path,
-            run_id="run-port-api",
-            name="port-api",
-            base_url_vars={"API_PORT": "443/path"},
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert 'base URL variable "API_PORT"' in invalid_vm.message
-
-    def test_credentialed_builtin_firewall_entry_rejects_http_dynamic_base(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-strapi",
-                            "firewalls": [
-                                {
-                                    "kind": "builtin",
-                                    "name": "strapi",
-                                    "baseUrlVars": {
-                                        "STRAPI_BASE_URL": "http://strapi.example.test"
-                                    },
-                                }
-                            ],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        invalid_vm = state.invalid_vms["10.200.0.1"]
-        assert invalid_vm.reason == "invalid_firewalls"
-        assert "credentialed base URL must use https" in invalid_vm.message
-
-    def test_credentialed_builtin_firewall_entry_accepts_https_dynamic_base(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-strapi",
-                            "firewalls": [
-                                {
-                                    "kind": "builtin",
-                                    "name": "strapi",
-                                    "baseUrlVars": {
-                                        "STRAPI_BASE_URL": "https://strapi.example.test"
-                                    },
-                                }
-                            ],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        assert vm_info["firewalls"][0]["apis"][0]["base"] == ("https://strapi.example.test")
-
     def test_inline_firewall_entry_allows_http_base(self, tmp_path):
         path = tmp_path / "registry.json"
         write_firewall_registry(path)
@@ -1138,82 +121,6 @@ class TestGetVmContext:
             matching.FirewallAllow,
         )
 
-    def test_builtin_firewall_entry_missing_dynamic_var_rejects_vm(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-zendesk",
-                            "firewalls": [{"kind": "builtin", "name": "zendesk"}],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-        assert "ZENDESK_SUBDOMAIN" in state.invalid_vms["10.200.0.1"].message
-
-    def test_builtin_firewall_entry_does_not_read_top_level_vars(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-zendesk",
-                            "vars": {"ZENDESK_SUBDOMAIN": "top-level"},
-                            "firewalls": [{"kind": "builtin", "name": "zendesk"}],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-        assert "ZENDESK_SUBDOMAIN" in state.invalid_vms["10.200.0.1"].message
-
-    def test_unknown_builtin_firewall_entry_rejects_vm(self, tmp_path):
-        path = tmp_path / "registry.json"
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-missing",
-                            "firewalls": [{"kind": "builtin", "name": "missing-firewall"}],
-                        }
-                    },
-                    "updatedAt": 0,
-                }
-            )
-        )
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-        assert "missing-firewall" in state.invalid_vms["10.200.0.1"].message
-
     def test_invalid_entry_has_no_context(self, tmp_path):
         path = tmp_path / "registry.json"
         path.write_text(
@@ -1235,251 +142,3 @@ class TestGetVmContext:
 
         assert not isinstance(state, registry.RegistryUnavailable)
         assert set(state.invalid_vms) == {"10.200.0.2"}
-
-    def test_compiled_context_updates_after_successful_registry_change(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path, rule="/items")
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        assert first_context is not None
-        first_vm_info, first_compiled, first_compiled_policies = first_context
-        assert first_compiled is not None
-
-        write_firewall_registry(path, rule="/other-resource")
-        second_context = registry.get_vm_context("10.200.0.1", str(path))
-        assert second_context is not None
-        second_vm_info, second_compiled, second_compiled_policies = second_context
-
-        assert second_vm_info is not first_vm_info
-        assert second_compiled is not None
-        assert second_compiled is not first_compiled
-        assert second_compiled_policies is not first_compiled_policies
-        assert isinstance(
-            matching.match_compiled_firewall_request(
-                "https://api.example.com/items",
-                "GET",
-                second_compiled,
-                second_compiled_policies,
-            ),
-            matching.FirewallBlock,
-        )
-        assert isinstance(
-            matching.match_compiled_firewall_request(
-                "https://api.example.com/other-resource",
-                "GET",
-                second_compiled,
-                second_compiled_policies,
-            ),
-            matching.FirewallAllow,
-        )
-
-    def test_successful_registry_change_without_firewalls_clears_compiled_context(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        first_context = registry.get_vm_context("10.200.0.1", str(path))
-        assert first_context is not None
-        _, first_compiled, first_compiled_policies = first_context
-        assert first_compiled is not None
-
-        path.write_text(
-            json.dumps(
-                {
-                    "vms": {
-                        "10.200.0.1": {
-                            "runId": "run-abc-123",
-                            "networkPolicies": {
-                                "example": {
-                                    "allow": [],
-                                    "deny": [],
-                                    "unknownPolicy": "allow",
-                                }
-                            },
-                        }
-                    },
-                    "updatedAt": 1700000000001,
-                }
-            )
-        )
-
-        second_context = registry.get_vm_context("10.200.0.1", str(path))
-        assert second_context is not None
-        _, second_compiled, second_compiled_policies = second_context
-        assert second_compiled is None
-        assert second_compiled_policies is not first_compiled_policies
-
-    def test_compiled_context_is_scoped_to_registry_path(self, tmp_path):
-        path_a = tmp_path / "registry-a.json"
-        path_b = tmp_path / "registry-b.json"
-        write_firewall_registry(path_a, rule="/items")
-        write_firewall_registry(path_b, rule="/other")
-        pin_mtime(path_a)
-        pin_mtime(path_b)
-        assert path_a.stat().st_size == path_b.stat().st_size
-
-        first_context = registry.get_vm_context("10.200.0.1", str(path_a))
-        second_context = registry.get_vm_context("10.200.0.1", str(path_b))
-
-        assert first_context is not None
-        assert second_context is not None
-        first_vm_info, first_compiled, _ = first_context
-        second_vm_info, second_compiled, second_compiled_policies = second_context
-        assert first_vm_info is not second_vm_info
-        assert second_compiled is not None
-        assert second_compiled is not first_compiled
-        assert isinstance(
-            matching.match_compiled_firewall_request(
-                "https://api.example.com/items",
-                "GET",
-                second_compiled,
-                second_compiled_policies,
-            ),
-            matching.FirewallBlock,
-        )
-        assert isinstance(
-            matching.match_compiled_firewall_request(
-                "https://api.example.com/other",
-                "GET",
-                second_compiled,
-                second_compiled_policies,
-            ),
-            matching.FirewallAllow,
-        )
-
-    def test_parse_failure_returns_no_compiled_context(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        context = registry.get_vm_context("10.200.0.1", str(path))
-        assert context is not None
-        _, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-
-        path.write_text("{ broken")
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            unavailable_context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert unavailable_context is None
-        assert isinstance(state, registry.RegistryUnavailable)
-        assert state.reason == "parse_failed"
-
-    def test_missing_file_returns_no_compiled_context(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        context = registry.get_vm_context("10.200.0.1", str(path))
-        assert context is not None
-        _, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-
-        path.unlink()
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            unavailable_context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert unavailable_context is None
-        assert isinstance(state, registry.RegistryUnavailable)
-        assert state.reason == "stat_failed"
-
-    def test_malformed_network_policy_shape_compiles_without_load_failure(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        data = json.loads(path.read_text())
-        data["vms"]["10.200.0.1"]["networkPolicies"] = {"example": "denied"}
-        path.write_text(json.dumps(data))
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        _, compiled_firewalls, compiled_network_policies = context
-        assert compiled_firewalls is not None
-        result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            compiled_firewalls,
-            compiled_network_policies,
-        )
-        assert isinstance(result, matching.FirewallBlock)
-        assert result.reason == "malformed_network_policy"
-
-    def test_malformed_firewall_config_compiles_without_load_failure(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        data = json.loads(path.read_text())
-        firewall = data["vms"]["10.200.0.1"]["firewalls"][0]["firewall"]
-        firewall["apis"][0]["permissions"][0]["rules"] = ["GET /items/{a}literal{b}"]
-        path.write_text(json.dumps(data))
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        _, compiled_firewalls, compiled_network_policies = context
-        assert compiled_firewalls is not None
-        result = matching.match_compiled_firewall_request(
-            "https://api.example.com/items",
-            "GET",
-            compiled_firewalls,
-            compiled_network_policies,
-        )
-        assert isinstance(result, matching.FirewallBlock)
-        assert result.reason == "malformed_firewall_config"
-
-    @pytest.mark.parametrize(
-        "firewalls",
-        [0, 1, False, True, "", {}, {"name": "example"}, "broken"],
-    )
-    def test_malformed_top_level_firewalls_shape_rejects_vm(self, tmp_path, firewalls):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        data = json.loads(path.read_text())
-        data["vms"]["10.200.0.1"]["firewalls"] = firewalls
-        path.write_text(json.dumps(data))
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert "10.200.0.1" not in state.vms
-        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-
-    @pytest.mark.parametrize(
-        "firewalls",
-        [
-            [0],
-            [{"kind": "inline"}],
-            [{"kind": "builtin", "name": ""}],
-            [{"kind": "builtin", "name": "zendesk", "baseUrlVars": []}],
-            [{"kind": "builtin", "name": "zendesk", "baseUrlVars": {"ZENDESK_SUBDOMAIN": 1}}],
-            [{"name": "github", "apis": []}],
-            [{"kind": "unknown", "name": "github"}],
-            [{"kind": "unknown", "name": "github", "apis": []}],
-        ],
-    )
-    def test_malformed_firewall_entries_reject_vm(self, tmp_path, firewalls):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        data = json.loads(path.read_text())
-        data["vms"]["10.200.0.1"]["firewalls"] = firewalls
-        path.write_text(json.dumps(data))
-
-        with patch.object(registry.ctx, "log", MagicMock(), create=True):
-            context = registry.get_vm_context("10.200.0.1", str(path))
-            state = registry.load_registry_state(str(path))
-
-        assert context is None
-        assert not isinstance(state, registry.RegistryUnavailable)
-        assert "10.200.0.1" not in state.vms
-        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
-
-    def test_null_top_level_firewalls_shape_is_no_firewall_context(self, tmp_path):
-        path = tmp_path / "registry.json"
-        write_firewall_registry(path)
-        data = json.loads(path.read_text())
-        data["vms"]["10.200.0.1"]["firewalls"] = None
-        path.write_text(json.dumps(data))
-
-        context = registry.get_vm_context("10.200.0.1", str(path))
-
-        assert context is not None
-        _, compiled_firewalls, compiled_network_policies = context
-        assert compiled_firewalls is None
-        assert compiled_network_policies is not None
