+++ crates/runner/mitm-addon/tests/test_registry_context_state.py
@@ -1,0 +1,260 @@
+"""Tests for registry compiled context state and reload behavior."""
+
+import json
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+import matching
+import registry
+from tests.registry_helpers import pin_mtime, write_firewall_registry
+
+
+class TestRegistryContextState:
+    def test_compiled_context_updates_after_successful_registry_change(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path, rule="/items")
+        first_context = registry.get_vm_context("10.200.0.1", str(path))
+        assert first_context is not None
+        first_vm_info, first_compiled, first_compiled_policies = first_context
+        assert first_compiled is not None
+
+        write_firewall_registry(path, rule="/other-resource")
+        second_context = registry.get_vm_context("10.200.0.1", str(path))
+        assert second_context is not None
+        second_vm_info, second_compiled, second_compiled_policies = second_context
+
+        assert second_vm_info is not first_vm_info
+        assert second_compiled is not None
+        assert second_compiled is not first_compiled
+        assert second_compiled_policies is not first_compiled_policies
+        assert isinstance(
+            matching.match_compiled_firewall_request(
+                "https://api.example.com/items",
+                "GET",
+                second_compiled,
+                second_compiled_policies,
+            ),
+            matching.FirewallBlock,
+        )
+        assert isinstance(
+            matching.match_compiled_firewall_request(
+                "https://api.example.com/other-resource",
+                "GET",
+                second_compiled,
+                second_compiled_policies,
+            ),
+            matching.FirewallAllow,
+        )
+
+    def test_successful_registry_change_without_firewalls_clears_compiled_context(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        first_context = registry.get_vm_context("10.200.0.1", str(path))
+        assert first_context is not None
+        _, first_compiled, first_compiled_policies = first_context
+        assert first_compiled is not None
+
+        path.write_text(
+            json.dumps(
+                {
+                    "vms": {
+                        "10.200.0.1": {
+                            "runId": "run-abc-123",
+                            "networkPolicies": {
+                                "example": {
+                                    "allow": [],
+                                    "deny": [],
+                                    "unknownPolicy": "allow",
+                                }
+                            },
+                        }
+                    },
+                    "updatedAt": 1700000000001,
+                }
+            )
+        )
+
+        second_context = registry.get_vm_context("10.200.0.1", str(path))
+        assert second_context is not None
+        _, second_compiled, second_compiled_policies = second_context
+        assert second_compiled is None
+        assert second_compiled_policies is not first_compiled_policies
+
+    def test_compiled_context_is_scoped_to_registry_path(self, tmp_path):
+        path_a = tmp_path / "registry-a.json"
+        path_b = tmp_path / "registry-b.json"
+        write_firewall_registry(path_a, rule="/items")
+        write_firewall_registry(path_b, rule="/other")
+        pin_mtime(path_a)
+        pin_mtime(path_b)
+        assert path_a.stat().st_size == path_b.stat().st_size
+
+        first_context = registry.get_vm_context("10.200.0.1", str(path_a))
+        second_context = registry.get_vm_context("10.200.0.1", str(path_b))
+
+        assert first_context is not None
+        assert second_context is not None
+        first_vm_info, first_compiled, _ = first_context
+        second_vm_info, second_compiled, second_compiled_policies = second_context
+        assert first_vm_info is not second_vm_info
+        assert second_compiled is not None
+        assert second_compiled is not first_compiled
+        assert isinstance(
+            matching.match_compiled_firewall_request(
+                "https://api.example.com/items",
+                "GET",
+                second_compiled,
+                second_compiled_policies,
+            ),
+            matching.FirewallBlock,
+        )
+        assert isinstance(
+            matching.match_compiled_firewall_request(
+                "https://api.example.com/other",
+                "GET",
+                second_compiled,
+                second_compiled_policies,
+            ),
+            matching.FirewallAllow,
+        )
+
+    def test_parse_failure_returns_no_compiled_context(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        context = registry.get_vm_context("10.200.0.1", str(path))
+        assert context is not None
+        _, compiled_firewalls, _ = context
+        assert compiled_firewalls is not None
+
+        path.write_text("{ broken")
+        with patch.object(registry.ctx, "log", MagicMock(), create=True):
+            unavailable_context = registry.get_vm_context("10.200.0.1", str(path))
+            state = registry.load_registry_state(str(path))
+
+        assert unavailable_context is None
+        assert isinstance(state, registry.RegistryUnavailable)
+        assert state.reason == "parse_failed"
+
+    def test_missing_file_returns_no_compiled_context(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        context = registry.get_vm_context("10.200.0.1", str(path))
+        assert context is not None
+        _, compiled_firewalls, _ = context
+        assert compiled_firewalls is not None
+
+        path.unlink()
+        with patch.object(registry.ctx, "log", MagicMock(), create=True):
+            unavailable_context = registry.get_vm_context("10.200.0.1", str(path))
+            state = registry.load_registry_state(str(path))
+
+        assert unavailable_context is None
+        assert isinstance(state, registry.RegistryUnavailable)
+        assert state.reason == "stat_failed"
+
+    def test_malformed_network_policy_shape_compiles_without_load_failure(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        data = json.loads(path.read_text())
+        data["vms"]["10.200.0.1"]["networkPolicies"] = {"example": "denied"}
+        path.write_text(json.dumps(data))
+
+        context = registry.get_vm_context("10.200.0.1", str(path))
+
+        assert context is not None
+        _, compiled_firewalls, compiled_network_policies = context
+        assert compiled_firewalls is not None
+        result = matching.match_compiled_firewall_request(
+            "https://api.example.com/items",
+            "GET",
+            compiled_firewalls,
+            compiled_network_policies,
+        )
+        assert isinstance(result, matching.FirewallBlock)
+        assert result.reason == "malformed_network_policy"
+
+    def test_malformed_firewall_config_compiles_without_load_failure(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        data = json.loads(path.read_text())
+        firewall = data["vms"]["10.200.0.1"]["firewalls"][0]["firewall"]
+        firewall["apis"][0]["permissions"][0]["rules"] = ["GET /items/{a}literal{b}"]
+        path.write_text(json.dumps(data))
+
+        context = registry.get_vm_context("10.200.0.1", str(path))
+
+        assert context is not None
+        _, compiled_firewalls, compiled_network_policies = context
+        assert compiled_firewalls is not None
+        result = matching.match_compiled_firewall_request(
+            "https://api.example.com/items",
+            "GET",
+            compiled_firewalls,
+            compiled_network_policies,
+        )
+        assert isinstance(result, matching.FirewallBlock)
+        assert result.reason == "malformed_firewall_config"
+
+    @pytest.mark.parametrize(
+        "firewalls",
+        [0, 1, False, True, "", {}, {"name": "example"}, "broken"],
+    )
+    def test_malformed_top_level_firewalls_shape_rejects_vm(self, tmp_path, firewalls):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        data = json.loads(path.read_text())
+        data["vms"]["10.200.0.1"]["firewalls"] = firewalls
+        path.write_text(json.dumps(data))
+
+        with patch.object(registry.ctx, "log", MagicMock(), create=True):
+            context = registry.get_vm_context("10.200.0.1", str(path))
+            state = registry.load_registry_state(str(path))
+
+        assert context is None
+        assert not isinstance(state, registry.RegistryUnavailable)
+        assert "10.200.0.1" not in state.vms
+        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
+
+    @pytest.mark.parametrize(
+        "firewalls",
+        [
+            [0],
+            [{"kind": "inline"}],
+            [{"kind": "builtin", "name": ""}],
+            [{"kind": "builtin", "name": "zendesk", "baseUrlVars": []}],
+            [{"kind": "builtin", "name": "zendesk", "baseUrlVars": {"ZENDESK_SUBDOMAIN": 1}}],
+            [{"name": "github", "apis": []}],
+            [{"kind": "unknown", "name": "github"}],
+            [{"kind": "unknown", "name": "github", "apis": []}],
+        ],
+    )
+    def test_malformed_firewall_entries_reject_vm(self, tmp_path, firewalls):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        data = json.loads(path.read_text())
+        data["vms"]["10.200.0.1"]["firewalls"] = firewalls
+        path.write_text(json.dumps(data))
+
+        with patch.object(registry.ctx, "log", MagicMock(), create=True):
+            context = registry.get_vm_context("10.200.0.1", str(path))
+            state = registry.load_registry_state(str(path))
+
+        assert context is None
+        assert not isinstance(state, registry.RegistryUnavailable)
+        assert "10.200.0.1" not in state.vms
+        assert state.invalid_vms["10.200.0.1"].reason == "invalid_firewalls"
+
+    def test_null_top_level_firewalls_shape_is_no_firewall_context(self, tmp_path):
+        path = tmp_path / "registry.json"
+        write_firewall_registry(path)
+        data = json.loads(path.read_text())
+        data["vms"]["10.200.0.1"]["firewalls"] = None
+        path.write_text(json.dumps(data))
+
+        context = registry.get_vm_context("10.200.0.1", str(path))
+
+        assert context is not None
+        _, compiled_firewalls, compiled_network_policies = context
+        assert compiled_firewalls is None
+        assert compiled_network_policies is not None
